More int to NSInteger conversions
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/back/branches/gnustep_testplant_branch@40243 72102866-910b-0410-8b05-ffd578937521

+++ Source/x11/XGDragView.m
@@ -266,7 +266,7 @@ static	XGDragView	*sharedDragView = nil;
                     action: (NSDragOperation)action
                   position: (NSPoint)eventLocation
                  timestamp: (NSTimeInterval)time
-                  toWindow: (int)dWindowNumber
+                  toWindow: (NSInteger)dWindowNumber
 {
   gswindow_device_t *dragWindev = DRAGWINDEV;
